[SPARK-39163][SQL] Throw an exception w/ error class for an invalid bucket file
### What changes were proposed in this pull request?
In the PR, I propose to use the INVALID_BUCKET_FILE error classes for an invalid bucket file.

### Why are the changes needed?
Porting the executing errors for multiple rows from a subquery used as an expression to the new error framework should improve user experience with Spark SQL.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #36603 from panbingkun/SPARK-39163.

Authored-by: panbingkun <pbk1982@gmail.com>
